Move fixture function to utils file
